Add files via upload
The project is done on retina(fundus) images. The algorithm is based on finding the blood vessels and extracting them. In this way, we clearly see the problematic areas such as hemorrhages. If their numbers are big(e.g. greater than 5), then the eye has considerable amount hemorrhages, given that hemorrhages on the funduses are mostly caused by glucose level, then we can say that the eye has Diabetic Retinopty.